 requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.0
pydantic-settings==2.5.0

# Transcription
faster-whisper==1.0.3

# Audio Processing
ffmpeg-python==0.2.0

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Monitoring
prometheus-client==0.20.0